# contracts/openapi/v1.yaml
openapi: "3.1.0"
info:
  title: Agent Framework Orchestrator API
  version: "1.0.0"
  description: |
    REST API for the Agent Framework control plane orchestrator.
    Accepts natural language specifications, decomposes them into
    executable tasks, and coordinates specialized AI agents.
  contact:
    name: Agent Framework Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://agent-framework-dev.azurecontainerapps.io
    description: Development environment
paths:
...
  /api/v1/rewards:
    get:
      operationId: exportRewards
      summary: Export per-task reward records as NDJSON
      description: |
        Returns per-task reward records as NDJSON (one JSON object per line).
        Optionally filtered by planId. Only items with at least one reward signal
        are returned. Format is DPO-ready: includes task metadata, result, and
        all reward components.
      tags:
        - Rewards
      parameters:
        - name: planId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter rewards by plan ID
      responses:
        "200":
          description: NDJSON stream of reward records
          content:
            application/x-ndjson:
              schema:
                type: string
                description: |
                  One JSON object per line with fields: taskKey, workerType,
                  workerProfile, status, aggregatedReward, reviewScore,
                  processScore, rewardSources, result
...
  /api/v1/rewards/preference-pairs:
    get:
      operationId: exportPreferencePairs
      summary: Export DPO preference pairs as NDJSON
      description: |
        Returns DPO-format preference pairs as NDJSON. Each pair contains
        a chosen (higher reward) and rejected (lower reward) response for the
        same task type, suitable for offline preference learning.
      tags:
        - Rewards
      parameters:
        - name: minDelta
          in: query
          required: false
          schema:
            type: number
            format: float
            default: 0.3
          description: Minimum reward delta for inclusion
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 500
          description: Maximum number of pairs to return
      responses:
        "200":
          description: NDJSON stream of preference pairs
          content:
            application/x-ndjson:
              schema:
                type: string
                description: |
                  One JSON object per line with fields: id, planId, taskKey,
                  workerType, prompt, chosen, rejected, chosenReward,
                  rejectedReward, deltaReward, generationSource, createdAt